requirements.txt
certifi==2018.11.29
Django==2.1.5
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-geoposition==0.3.0
django-summernote==0.8.11.4
django-taggit==0.23.0
djangorestframework==3.9.1
nodeenv==1.3.3
Pillow==5.4.1
pytz==2018.9
sorl-thumbnail==12.5.0
